pipelines/fast_imaging_detection_pipeline_distributed_flink/development/requirements.txt
numpy
astropy
matplotlib
pybase64
python-casacore
natsort
Pillow
bdsf
scipy
smart_open